refactor: simplify event screens
