+++ packages/apps/crm/src/components/organisms/Event/index.js
@@ -1,0 +1,6 @@
+export {default as EventContactCard} from './EventContactCard/EventContactCard';
+export {default as EventDatesCard} from './EventDatesCard/EventDatesCard';
+export {default as EventHeader} from './EventHeader/EventHeader';
+export {default as EventLabelsCard} from './EventLabelsCard/EventLabelsCard';
+export {default as EventLeadCard} from './EventLeadCard/EventLeadCard';
+export {default as EventPartnerCard} from './EventPartnerCard/EventPartnerCard';
