+++ packages/apps/crm/src/components/organisms/index.js
@@ -1,23 +1,12 @@
-export {default as ClientBody} from './ClientBody/ClientBody';
-export {default as ClientBottom} from './ClientBottom/ClientBottom';
-export {default as ClientHeader} from './ClientHeader/ClientHeader';
-export {default as ContactBody} from './ContactBody/ContactBody';
-export {default as ContactBottom} from './ContactBottom/ContactBottom';
-export {default as ContactHeader} from './ContactHeader/ContactHeader';
 export {default as DropdownContactView} from './DropdownContactView/DropdownContactView';
-export {default as DropdownGeneralView} from './DropdownGeneralView/DropdownGeneralView';
-export {default as EditButton} from './EditButton/EditButton';
 export {default as DropdownEmployeeView} from './DropdownEmployeeView/DropdownEmployeeView';
 export {default as DropdownEventView} from './DropdownEventView/DropdownEventView';
-export {default as EventHeader} from './EventHeader/EventHeader';
-export {default as EventBody} from './EventBody/EventBody';
-export {default as LeadBody} from './LeadBody/LeadBody';
-export {default as LeadBottom} from './LeadBottom/LeadBottom';
-export {default as LeadHeader} from './LeadHeader/LeadHeader';
-export {default as OpportunityBody} from './OpportunityBody/OpportunityBody';
-export {default as OpportunityBottom} from './OpportunityBottom/OpportunityBottom';
-export {default as OpportunityDropdownInfo} from './OpportunityDropdownInfo/OpportunityDropdownInfo';
-export {default as OpportunityHeader} from './OpportunityHeader/OpportunityHeader';
-export {default as ProspectBody} from './ProspectBody/ProspectBody';
-export {default as ProspectBottom} from './ProspectBottom/ProspectBottom';
-export {default as ProspectHeader} from './ProspectHeader/ProspectHeader';
+export {default as DropdownGeneralView} from './DropdownGeneralView/DropdownGeneralView';
+export {default as EditButton} from './EditButton/EditButton';
+
+export * from './Client';
+export * from './Contact';
+export * from './Event';
+export * from './Lead';
+export * from './Opportunity';
+export * from './Prospect';
