+++ packages/apps/crm/src/screens/event/EventDetailsScreen.js
@@ -1,15 +1,20 @@
 import React, {useEffect} from 'react';
 import {ScrollView} from 'react-native';
-import {Screen, HeaderContainer} from '@axelor/aos-mobile-ui';
-import {useDispatch, useSelector} from '@axelor/aos-mobile-core';
+import {Screen, HeaderContainer, NotesCard} from '@axelor/aos-mobile-ui';
+import {useDispatch, useSelector, useTranslator} from '@axelor/aos-mobile-core';
 import {fetchEventById} from '../../features/eventSlice';
-import {fetchLeadById, fetchLeadStatus} from '../../features/leadSlice';
-import {fetchPartner} from '../../features/partnerSlice';
-import {getContact} from '../../features/contactSlice';
-import {EventBody, EventHeader} from '../../components';
+import {
+  EventContactCard,
+  EventDatesCard,
+  EventHeader,
+  EventLabelsCard,
+  EventLeadCard,
+  EventPartnerCard,
+} from '../../components';
 
 function EventDetailsScreen({navigation, route}) {
   const eventId = route.params.eventId;
+  const I18n = useTranslator();
   const dispatch = useDispatch();
 
   const {event} = useSelector(state => state.event);
@@ -17,20 +22,6 @@ function EventDetailsScreen({navigation, route}) {
   useEffect(() => {
     dispatch(fetchEventById({eventId: eventId}));
   }, [dispatch, eventId]);
-
-  useEffect(() => {
-    event.lead && dispatch(fetchLeadById({leadId: event.lead.id}));
-    event.lead && dispatch(fetchLeadStatus());
-  }, [dispatch, event.lead]);
-
-  useEffect(() => {
-    event?.partner && dispatch(fetchPartner({partnerId: event?.partner?.id}));
-  }, [dispatch, event?.partner]);
-
-  useEffect(() => {
-    event?.contactPartner &&
-      dispatch(getContact({contactId: event?.contactPartner.id}));
-  }, [dispatch, event?.contactPartner]);
 
   if (event?.id !== eventId) {
     return null;
@@ -40,7 +31,12 @@ function EventDetailsScreen({navigation, route}) {
     <Screen removeSpaceOnTop={true}>
       <HeaderContainer expandableFilter={false} fixedItems={<EventHeader />} />
       <ScrollView>
-        <EventBody navigation={navigation} />
+        <EventDatesCard />
+        <EventLabelsCard />
+        <EventLeadCard navigation={navigation} />
+        <EventPartnerCard navigation={navigation} />
+        <EventContactCard navigation={navigation} />
+        <NotesCard title={I18n.t('Crm_Description')} data={event.description} />
       </ScrollView>
     </Screen>
   );
